msvc warnings fixes
darcs-hash:20061009211509-16063-7c8f72f49b3f4255583fb0e223c377bd4a1d23be.gz

+++ libsofia-sip-ua/sip/sip_tag_class.c
@@ -422,7 +422,7 @@ tagi_t *sip_url_query_as_taglist(su_home_t *home, char const *query,
       hc = (msg_hclass_t *)sip_tag_list[j]->tt_magic;
       if (n == 1 && strncasecmp(hnv, hc->hc_short, 1) == 0)
 	break;
-      else if (n == hc->hc_len && strncasecmp(hnv, hc->hc_name, n) == 0)
+      else if (n == (size_t)hc->hc_len && strncasecmp(hnv, hc->hc_name, n) == 0)
 	break;
     }
